pyproject.toml
[tool.poetry]
name = "dio-python-twitter-py"
version = "0.1.0"
description = "Integração com twitter"
authors = ["Diego Rodrigues"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^3.12.0"
uvicorn = "^0.15.0"
python-dotenv = "^1.0.0"
tweepy = "^3.10.0"
fastapi = "^0.68.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
